Add special condition for hexagonal lattices.
According to page 20 of T. Fritz, Molecular Architecture in Heteroepitaxially Grown Organic Thin Films one needs to treat hexagonal lattices specially, as for each sets of lattice vectors a1,a2 there exists a fully equivalent set given by a1'=a1, and a2'=a1-a2.
The latest change of the code checks if α is exactly a multiple of 60 degrees (except for the clearly nonsensical 180 and 0 degrees) and if that's the case, runs the main loop twice, with alpha set to 60° and 120°. By this, the mirror symmetry of the hexagonal lattice is preserved.